+++ docs/make.jl
@@ -30,6 +30,7 @@ makedocs(;
             "guide/parallel_tempering.md",
             "guide/ground_states.md",
             "guide/parallelism.md",
+            "guide/gpu.md",
             "guide/observables.md",
             "guide/checkpointing.md",
         ],